pyproject.toml
[tool.poetry]
name = "hackathon-backend"
version = "0.1.0"
description = "hackathon-backend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
SQLAlchemy = {extras = ["async"], version = "^1.4.41"}
uvicorn = {extras = ["standard"], version = "^0.18.3"}
alembic = "^1.8.1"
asyncpg = "^0.26.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
black = "^22.8.0"
coverage = "^6.4.4"
isort = "^5.10.1"
psycopg2 = "^2.9.3"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.19.0"
pytest-mock = "^3.9.0"
pytest-xdist = "^2.5.0"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.env
  | \.venv
  | venv
  | env
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 119
lines_after_imports = 2
multi_line_output = 3
force_grid_wrap = 2
skip_glob = ["**/env/*"]

[tool.pytest.ini_options]
addopts = "-v --asyncio-mode auto --cov=."
testpaths = ["tests"]
filterwarnings = [
    "ignore::sqlalchemy.exc.SAWarning"
]
markers = [
    "integration: run integration tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
